requirements.txt
django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
whitenoise==6.0.0
django-heroku==0.3.1
gunicorn==20.1.0
django-filter